start refactor, some commands not working.
