+++ index.js
@@ -4,14 +4,9 @@ const client = new Discord.Client();
 var GphApiClient = require('giphy-js-sdk-core')
 giphyclient = GphApiClient(process.env.GIPHYAPIKEY)
 const prefix = "bawt";
-const version = "yo";
 const keywords = require('./keywords');
-const help = require('./help');
 var CronJob = require('cron').CronJob;
-const redditQuote = require ('./reddit_quote')
-const googleSuggest = require ('./google_suggest')
-const meme = require ('./meme')
-const defineWord = require ('./google-define')
+const botMessage = require ('./bot_message')
 var Jimp = require('jimp');
 var mainChannel;
 var knex = require('knex')({
@@ -44,13 +39,6 @@ client.on("ready", () => {
 
 const keyWords = keywords.keyWords();
 const keyWordsJson = keywords.keyWordsJson();
-
-const roll = (message) => {
-  let min = 1;
-  let max = 100;
-  let roll = Math.floor(Math.random() * (max - min) + 1);
-  return `${message.author.username} rolled ${roll} (${min}-${max})`
-}
 
 const initializeNewsWatcher = (channel) => {
   var Watcher  = require('feed-watcher'),
@@ -214,120 +202,8 @@ client.on("message", (message) => {
 
   const args = message.content.slice(prefix.length).trim().split(" ");
   const command = args.shift().toLowerCase().trim();
-
-  if (command === "sup") {
-    message.channel.send(version);
-  }
-  else if (command === "avatar") {
-    if (!message.mentions.users.size) {
-        return message.channel.send(message.author.displayAvatarURL);
-    }
-    const avatarList = message.mentions.users.map(user => {
-        return user.displayAvatarURL;
-    });
-    message.channel.send(avatarList);
-  }
-  else if (command === "roll") {
-    message.channel.send(roll(message))
-  }
-  else if (command === "gif")
-  {
-    let query = args.join(' ');
-    message.channel.send(query);
-    let rating = "pg-13";
-    if(message.channel.nsfw)
-    {
-      rating = "r";
-    }
-    giphyclient.search('gifs', {"q": query, "limit" : 10, "rating" : rating})
-    .then((response) => {
-      message.channel.send(response.data[Math.floor(Math.random() * (10 - 1) + 1)].url);
-    })
-    .catch((err) => {
-      message.channel.send("giphy didn't work....");
-    })
-  }
-  else if (command === "hot-gif")
-  {
-    let rating = "pg-13";
-    giphyclient.trending("gifs", {"rating" : rating})
-    .then((response) => {
-      message.channel.send(response.data[Math.floor(Math.random() * (10 - 1) + 1)].url);
-    })
-    .catch((err) => {
-      console.log(err);
-    })
-  }
-  else if (command === 'trump-quote') {
-    fetch("https://api.whatdoestrumpthink.com/api/v1/quotes/random")
-    .then(function(response) {
-        return response.json();
-      })
-      .then(function(trumpJson) {
-        message.channel.send(trumpJson.message);
-    });
-  }
-  else if (command === 'lukas-quote') {
-    redditQuote.new().then(function(response) {
-      message.channel.send(response);
-    })
-  }
-  else if (command === 'opinion') {
-    let query = args.join(' ');
-    googleSuggest.new(query).then(function(response) {
-      if (response) {
-        message.channel.send(response);
-      } else {
-        message.channel.send('Try google, bruv : |')
-      }
-    })
-  }
-  else if (command === 'dank-meme') {
-    meme.new('dankmemes').then(function(meme) {
-      message.channel.send({embed: meme})
-    })
-  }
-  else if (command === 'trump-meme') {
-    meme.new('presidenttrumptwitter').then(function(meme) {
-      message.channel.send({embed: meme})
-    })
-  }
-  else if (command === 'crackhead-cl') {
-    meme.new('crackheadcraigslist').then(function(meme) {
-      message.channel.send({embed: meme})
-    })
-  }
-  else if (command === 'stockphoto') {
-    meme.new('shittystockphotos').then(function(meme) {
-      message.channel.send({embed: meme})
-    })
-  }
-  else if (command === 'bpt-meme') {
-    meme.new('blackpeopletwitter').then(function(meme) {
-      message.channel.send({embed: meme})
-    })
-  }
-  else if (command === 'starterpack') {
-    meme.new('starterpacks').then(function(meme) {
-      message.channel.send({embed: meme})
-    })
-  }
-  else if (command === 'comic') {
-    message.channel.send(`generated comic`, {
-      files: ["./comic-.png"]
-    });
-  }
-  else if (command === 'help') {
-    message.channel.send(help.commands());
-  }
-  else if (command === 'define') {
-    word = args[0]
-    typeOfWord = args[1]
-    let query = {word: word, typeOfWord: typeOfWord}
-    defineWord.new(query).then(function(definition){
-      response = `${query.word}: ${definition}`
-      message.channel.send(response)
-    })
+  if (command) {
+    message.channel.send(botMessage.new(command, message, args))
   }
 });
